backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
PyMuPDF>=1.24.0
google-generativeai>=0.7.0
python-dotenv>=1.0.1
Werkzeug>=3.0.0
